Verify communication between client and server
Simple communication between client and server. The client waits for
user input, pack it into a json object of the correct format and send it
to server. If the server receives it correctly it will respond with a
message of type 'success' with a timestamp. 'Sender' and 'Content'
fields are left blank for now. The client receives the success response
and notify user.

Proper disconnection, more advanced server and client handling (be able
to handle more responses and multiple users) is not implemented.