Parser: Nicer format for long messages
Long messages will now (hopefully) have a nicer format when displayed to
user. Previously the string would wrap at the end of the terminal and
kind of "overlap" with the info column (time and names). It didn't look
nice, and it felt out of place.

Now if the text is too long, it will wrap down to the next ':' sign,
below the first, and continue from there. The width is set to ~80
characters.

Also added some minor cosmetic changes.